Introduce PUID and PGID env variables to syncserver Dockerfile
This gives more flexibility to solve permission problems with volumes.

When mounting volumes in Docker, file ownership and permissions need to match
between the container and the host system to avoid permission issues. By
specifying PUID and PGID, you can ensure the container operates with the same
user and group privileges as the host user.